chore(defender): stop using connext sdk
