+++ packages/defender/package.json
@@ -24,8 +24,9 @@
     "typescript": "^4.9.5"
   },
   "dependencies": {
-    "@connext/sdk": "^0.2.7-alpha.3",
+    "axios": "^1.3.4",
     "defender-autotask-utils": "^1.39.0",
-    "defender-relay-client": "^1.39.0"
+    "defender-relay-client": "^1.39.0",
+    "ethers": "^6.0.8"
   }
 }
