+++ packages/defender/tsup.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  name: "tsup",
+  treeshake: true,
+  entry: ['autotasks'],
+  outExtension() {
+    return {
+      js: `.js`,
+    }
+  },
+});
